refactor: Extract datasource files
