+++ app/components/data-source-menu.tsx
@@ -1,115 +1,21 @@
 import { Trans } from "@lingui/macro";
 import { Typography } from "@mui/material";
-import { keyBy } from "lodash";
 import { useRouter } from "next/router";
-import { createContext, Dispatch, ReactNode, useContext, useMemo } from "react";
+import { ReactNode, useMemo } from "react";
 import React from "react";
 
 import Flex from "@/components/flex";
 import { Label, MinimalisticSelect } from "@/components/form";
-import { DataSource } from "@/configurator";
-import { WHITELISTED_DATA_SOURCES } from "@/domain/env";
 import {
+  DataSourceStateContext,
   stringifyDataSource,
-  retrieveDataSourceFromLocalStorage,
-  saveDataSourceToLocalStorage,
-  parseDataSource,
-} from "@/graphql/resolvers/data-source";
-import useEvent from "@/lib/use-event";
-import { useRouteState } from "@/lib/use-sync-router-param";
-import { DEFAULT_DATA_SOURCE } from "@/rdf/sparql-client";
-
-export const SOURCE_OPTIONS = [
-  {
-    value: "sparql+https://lindas.admin.ch/query",
-    label: "Prod",
-    position: 3,
-    isTrusted: true,
-  },
-  {
-    value: "sparql+https://int.lindas.admin.ch/query",
-    label: "Int",
-    position: 2,
-    isTrusted: false,
-  },
-  {
-    value: "sparql+https://test.lindas.admin.ch/query",
-    label: "Test",
-    position: 1,
-    isTrusted: false,
-  },
-].filter((d) => WHITELISTED_DATA_SOURCES.includes(d.label));
-
-const SOURCES_BY_LABEL = keyBy(SOURCE_OPTIONS, (d) => d.label);
-const SOURCES_BY_VALUE = keyBy(SOURCE_OPTIONS, (d) => d.value);
-
-export const useIsTrustedDataSource = (dataSource: DataSource) => {
-  return useMemo(() => {
-    const stringifiedDataSource = stringifyDataSource(dataSource);
-    return SOURCES_BY_VALUE[stringifiedDataSource]?.isTrusted;
-  }, [dataSource]);
-};
-
-const DataSourceStateContext = createContext<
-  { dataSource: DataSource; setDataSource: Dispatch<string> } | undefined
->(undefined);
-
-export const useDataSource = () => {
-  const ctx = useContext(DataSourceStateContext);
-
-  if (ctx === undefined) {
-    throw Error(
-      "You need to wrap the application in <DataSourceProvider /> to useDataSource()"
-    );
-  }
-
-  return ctx;
-};
-
-const parseSourceByLabel = (label: string): DataSource | undefined => {
-  const newSource = SOURCES_BY_LABEL[label];
-  return newSource ? parseDataSource(newSource.value) : undefined;
-};
-
-const sourceToLabel = (source: DataSource) => {
-  return SOURCES_BY_VALUE[stringifyDataSource(source)]?.label;
-};
-
-const getURLParam = (param: string) => {
-  const url =
-    typeof window !== "undefined" ? new URL(window.location.href) : null;
-  if (!url) {
-    return undefined;
-  }
-  return url.searchParams.get(param);
-};
+  useDataSource,
+  useDataSourceState,
+} from "@/domain/datasource";
+import { SOURCE_OPTIONS } from "@/domain/datasource/constants";
 
 export const DataSourceProvider = ({ children }: { children: ReactNode }) => {
-  const [source, setSource] = useRouteState(
-    () => {
-      // Cannot use next router here as it is initialized asynchronously
-      const urlParam = getURLParam("dataSource");
-
-      // Priority for initial state: URL -> localStorage -> default
-      const initial =
-        (urlParam && parseSourceByLabel(urlParam)) ||
-        retrieveDataSourceFromLocalStorage() ||
-        DEFAULT_DATA_SOURCE;
-      return initial;
-    },
-    {
-      param: "dataSource",
-      deserialize: (l) => parseSourceByLabel(l) || DEFAULT_DATA_SOURCE,
-      serialize: sourceToLabel,
-      onValueChange: (newSource) => {
-        saveDataSourceToLocalStorage(newSource);
-      },
-    }
-  );
-
-  const setSourceByValue = useEvent((sourceValue: string) => {
-    setSource(parseDataSource(SOURCES_BY_VALUE[sourceValue]?.value));
-  });
+  const [source, setSourceByValue] = useDataSourceState();
 
   return (
     <DataSourceStateContext.Provider
