+++ app/domain/datasource/constants.ts
@@ -1,0 +1,27 @@
+import { keyBy } from "lodash";
+
+import { WHITELISTED_DATA_SOURCES } from "../env";
+
+export const SOURCE_OPTIONS = [
+  {
+    value: "sparql+https://lindas.admin.ch/query",
+    label: "Prod",
+    position: 3,
+    isTrusted: true,
+  },
+  {
+    value: "sparql+https://int.lindas.admin.ch/query",
+    label: "Int",
+    position: 2,
+    isTrusted: false,
+  },
+  {
+    value: "sparql+https://test.lindas.admin.ch/query",
+    label: "Test",
+    position: 1,
+    isTrusted: false,
+  },
+].filter((d) => WHITELISTED_DATA_SOURCES.includes(d.label));
+
+export const SOURCES_BY_LABEL = keyBy(SOURCE_OPTIONS, (d) => d.label);
+export const SOURCES_BY_VALUE = keyBy(SOURCE_OPTIONS, (d) => d.value);
